fix: change folder organisation
