+++ src/components/Board/BoardContainer.tsx
@@ -1,10 +1,12 @@
 import { connect } from 'react-redux'
 
 import Board from './Board'
-import { actionCreators } from '../../redux/boards/action'
-import { RootState, Dispatch } from '../../redux/RootReducers'
+import { actionCreators } from '../../redux/boards/actions'
+import { RootState, Dispatch } from '../../redux/RootReducer'
+
 import { IList } from '../../redux/lists/types'
 import { ITag } from '../../redux/tags/types'
+import { IUserRoleInBoard } from '../../redux/boards/types'
 
 const mapStateToProps = (state: RootState) => {
     return { board: state.board }
@@ -12,11 +14,19 @@ const mapStateToProps = (state: RootState) => {
 
 const mapDispatchToProps = (dispatch: Dispatch) => {
     return {
-        create: (id: number, title: string, isPrivate: boolean, lists: IList[], tags: ITag[]) => {
-            dispatch(actionCreators.createBoard(id, title, isPrivate, lists, tags))
+        create: (   id: number, 
+                    title: string, 
+                    isPrivate: boolean, 
+                    lists: IList[], 
+                    tags: ITag[], 
+                    userRole: IUserRoleInBoard[]) => {
+            dispatch(actionCreators.createBoardRequest(title, isPrivate, lists, tags, userRole))
         },
         delete: (index: number) => {
-            dispatch(actionCreators.removeBoard(index))
+            dispatch(actionCreators.removeBoardRequest(index))
+        },
+        setTitle: (title: string) => {
+            return // TODO for Alexis
         }
     }
 }
