+++ src/components/CardsList/DroppableCardsList.tsx
@@ -4,7 +4,7 @@ import { Droppable } from 'react-beautiful-dnd'
 import CardsList, { CardsListProps } from './CardsList'
 
 const DroppableCardsList: React.StatelessComponent<CardsListProps> = (props) => (
-    <Droppable droppableId={`droppable-${props.id}`} type="CARD">
+    <Droppable droppableId={`droppable-${props}`} type="CARD">
         {(provided, snapshot) => (
             <div ref={provided.innerRef}>
                 <CardsList {...props} />
