+++ src/redux/RootReducer.ts
@@ -2,8 +2,8 @@ import { combineReducers, Dispatch as ReduxDispatch } from 'redux'
 import { routerReducer as router, RouterState } from 'react-router-redux'
 import { reducer as lists, State as ListsState } from './lists/reducers'
 import { reducer as login, State as LoginState } from './login/reducers'
-import { reducer as boards, State as BoardsState } from './boards/reducers'
-import { reducer as board, State as BoardState } from './boards/reducer'
+import { reducer as boards, State as BoardsState } from './boards/boardsList/reducers'
+import { reducer as board, State as BoardState } from './boards/reducers'
 
 interface StoreEnhancerState { }
